Move contracts into subfolder
