+++ config/config.js
@@ -1,8 +1,8 @@
 import crypto from "@mainsail/core/bin/config/testnet/core/crypto.json" with { type: "json" };
 import DARK20 from "./DARK20.json" with { type: "json" };
-import { consensus } from "./consensus.js";
-import { usernames } from "./usernames.js";
-import { multiPayment } from "./multipayment.js";
+import { consensus } from "./contracts/consensus.js";
+import { usernames } from "./contracts/usernames.js";
+import { multiPayment } from "./contracts/multipayment.js";
 
 const config = {
     senderPassphrase: "", // REPLACE senderPassphrase WITH THE PASSPHRASE OF YOUR WALLET
